Migrate grade object to double
although it's not used but while parsing non int values crash the application

+++ app/src/main/java/trianglz/models/Assignments.java
@@ -22,7 +22,7 @@ public class Assignments {
 	@SerializedName("total")
 	public int total;
 	@SerializedName("grade")
-	public int grade;
+	public double grade;
 	@SerializedName("status")
 	public int status;
 	@SerializedName("hide_grade")
